requirements.txt
streamlit
pydantic
python-dotenv
requests
google-generativeai